obsidian-backup-master/requirements.txt
cffi==1.14.6
cryptography==3.4.7
pycparser==2.20
python-dotenv==0.19.0
